Refactor verifyToken
