+++ server/endpoints/AdminActions.ts
@@ -1,11 +1,10 @@
 import { body } from "express-validator";
 
+import { getCrossListOR, getMetricValues } from "common/CourseCard";
 import { Endpoint } from "../endpoints";
 import { Reviews, ReviewDocument, Classes } from "../dbDefs";
-import { verifyToken } from "./Auth";
 import { updateProfessors, findAllSemesters, resetProfessorArray } from "../dbInit"
-import { getCrossListOR, getMetricValues } from "../../common/CourseCard";
-import { getCourseById } from "./utils";
+import { getCourseById, verifyToken } from "./utils";
 
 // The type for a request with an admin action for a review
 interface AdminReviewRequest {
