+++ server/endpoints/utils.ts
@@ -1,5 +1,6 @@
 import { CourseId } from "./Review";
 import { Classes } from "../dbDefs";
+import { getUserByNetId, getVerificationTicket } from "./Auth";
 
 // eslint-disable-next-line import/prefer-default-export
 export const getCourseById = async (courseId: CourseId) => {
@@ -16,5 +17,27 @@ export const getCourseById = async (courseId: CourseId) => {
     // eslint-disable-next-line no-console
     console.log(error);
     return { error: "Internal Server Error" };
+  }
+};
+
+export const verifyToken = async (token: string) => {
+  try {
+    const regex = new RegExp(/^(?=.*[A-Z0-9])/i);
+    if (regex.test(token)) {
+      const ticket = await getVerificationTicket(token);
+      if (ticket && ticket.email) {
+        const user = await getUserByNetId(ticket.email.replace('@cornell.edu', ''));
+        if (user) {
+          return user.privilege === 'admin';
+        }
+      }
+    }
+    return false;
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.log("Error: at 'verufyToken' method");
+    // eslint-disable-next-line no-console
+    console.log(error);
+    return false;
   }
 };
